requirements.txt
pandas
gradio
google-generativeai
serpapi
sentence-transformers
torch
git+https://github.com/openai/CLIP.git
transformers
